Fully working ntfy urls and dynamic IP scanner
The basic modules for ntfy urls and dynamic ip scans entirely works. At this point I have most of the initial building done. Now time to bug squash and build up the rest of the modules. The linux health one might be the toughest. That one needs some real reworking