WFC Full implementation
I made cron work for scheduling, created a datatable to display current wfc scans that are setup, setup the continous task for it to coninue running on the interval that we configure, and setup the GUI to handle this. Also an abvility to delete them. The only thing remaining to implement is handling to scan results. Where they send to